backend/requirements.txt
Flask
flask-cors
playwright
requests
twilio
dotenv
openai
python-dotenv
retell-sdk
# Adding Bright Data MCP dependency
nodejs
